Add shimmer screen, delete confirmation, null checks, beautify code

+++ src/App.js
@@ -1,16 +1,17 @@
-import React, { useState, useEffect } from 'react';
-import ReactDOM  from 'react-dom/client';
-import axios from 'axios';
-import MessageForm from './components/MessageForm';
-import MessageList from './components/MessageList';
+import React, { useState, useEffect } from "react";
+import ReactDOM from "react-dom/client";
+import axios from "axios";
+import MessageForm from "./components/MessageForm";
+import MessageList from "./components/MessageList";
 
-const BASE_URL = 'https://mapi.harmoney.dev/api';
-const AUTH_TOKEN = 'FPZAggtFMinDlG1r';
+const BASE_URL = "https://mapi.harmoney.dev/api";
+const AUTH_TOKEN = "FPZAggtFMinDlG1r";
 
 function App() {
   const [messages, setMessages] = useState([]);
   const [sortAsc, setSortAsc] = useState(true);
   const [currentPage, setCurrentPage] = useState(1);
+  const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     fetchMessages(currentPage);
@@ -18,6 +19,7 @@ function App() {
 
   const fetchMessages = async (page = 1, pageSize = 10) => {
     try {
+      setLoading(true);
       const response = await axios.get(`${BASE_URL}/v1/messages/`, {
         headers: {
           Authorization: AUTH_TOKEN,
@@ -27,9 +29,11 @@ function App() {
           pageSize,
         },
       });
-      setMessages(response.data);
+      setMessages(response?.data);
     } catch (error) {
-      console.error('Error fetching messages:', error);
+      console.error("Error fetching messages:", error);
+    } finally {
+      setLoading(false);
     }
   };
 
@@ -43,14 +47,18 @@ function App() {
 
   const postMessage = async (newMessage) => {
     try {
-      await axios.post(`${BASE_URL}/v1/messages/`, { text: newMessage }, {
-        headers: {
-          Authorization: AUTH_TOKEN,
-        },
-      });
+      await axios.post(
+        `${BASE_URL}/v1/messages/`,
+        { text: newMessage },
+        {
+          headers: {
+            Authorization: AUTH_TOKEN,
+          },
+        }
+      );
       fetchMessages(); // Fetch messages immediately after posting
     } catch (error) {
-      console.error('Error posting message:', error);
+      console.error("Error posting message:", error);
     }
   };
 
@@ -61,27 +69,34 @@ function App() {
           Authorization: AUTH_TOKEN,
         },
       });
-      const updatedMessages = messages.filter((message) => message.id !== id);
+      const updatedMessages = messages?.filter((message) => message.id !== id);
       setMessages(updatedMessages);
     } catch (error) {
-      console.error('Error deleting message:', error);
+      console.error("Error deleting message:", error);
     }
   };
 
   const deleteAllMessages = async () => {
     try {
-      for (const message of messages) {
-        await axios.delete(`${BASE_URL}/v1/messages/${message.id}/`, {
-          headers: {
-            Authorization: AUTH_TOKEN,
-          },
-        });
+      const confirmDelete = window.confirm(
+        "Are you sure you want to delete all messages?"
+      );
+      if (confirmDelete) {
+        if (messages && messages.length > 0) {
+          for (const message of messages) {
+            await axios.delete(`${BASE_URL}/v1/messages/${message.id}/`, {
+              headers: {
+                Authorization: AUTH_TOKEN,
+              },
+            });
+          }
+        }
+
+        // Clear the messages state after deletion
+        setMessages([]);
       }
-  
-      // Clear the messages state after deletion
-      setMessages([]);
     } catch (error) {
-      console.error('Error deleting all messages:', error);
+      console.error("Error deleting all messages:", error);
     }
   };
 
@@ -92,26 +107,34 @@ function App() {
       return sortAsc ? timestampA - timestampB : timestampB - timestampA;
     });
 
-    setMessages(sortedMessages);
+    setMessages(sortedMessages || []);
     setSortAsc(!sortAsc);
   };
 
   return (
     <div className="app-container">
       <MessageForm onPostMessage={postMessage} />
-      <div>
-        <button onClick={deleteAllMessages}>Delete All Messages</button>
-        <button onClick={sortMessages}>Sort by Timestamp</button>
-        <MessageList messages={messages} onDeleteMessage={deleteMessage} />
-      </div>
-      <button onClick={prevPage} disabled={currentPage === 1}>Previous</button>
+
+      <button
+        onClick={deleteAllMessages}
+        disabled={!messages || messages.length === 0}
+      >
+        Delete All Messages
+      </button>
+      <button onClick={sortMessages}>Sort by Timestamp</button>
+      <MessageList
+        messages={messages}
+        onDeleteMessage={deleteMessage}
+        loading={loading}
+      />
+
+      <button onClick={prevPage} disabled={currentPage === 1}>
+        Previous
+      </button>
       <button onClick={nextPage}>Next</button>
     </div>
   );
 }
 
-
-
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(<App />);
-
